requirements.txt
absl-py==0.6.1
astor==0.7.1
atomicwrites==1.2.1
attrs==19.1.0
backcall==0.1.0
bleach==3.1.0
certifi==2019.3.9
cloudpickle==0.6.1
cycler==0.10.0
dask==1.0.0
decorator==4.3.0
defusedxml==0.5.0
entrypoints==0.3
future==0.17.1
gast==0.2.0
grpcio==1.16.1
h5py==2.8.0
imageio==2.4.1
ipdb==0.11
ipykernel==5.1.0
ipython==7.2.0
ipython-genutils==0.2.0
ipywidgets==7.4.2
jedi==0.13.2
Jinja2==2.10.1
jsonschema==3.0.1
jupyter==1.0.0
jupyter-client==5.2.4
jupyter-console==6.0.0
jupyter-core==4.4.0
Keras==2.2.4
Keras-Applications==1.0.6
Keras-Preprocessing==1.0.5
kiwisolver==1.0.1
Markdown==3.0.1
MarkupSafe==1.1.1
matplotlib==3.0.1
mistune==0.8.4
mkl-fft==1.0.6
mkl-random==1.0.1
more-itertools==4.3.0
nbconvert==5.4.1
nbformat==4.4.0
networkx==2.2
notebook==5.7.8
numpy==1.15.4
olefile==0.46
pandas==0.24.1
pandocfilters==1.4.2
parso==0.3.4
pexpect==4.6.0
pickleshare==0.7.5
Pillow==5.3.0
pluggy==0.8.0
prometheus-client==0.6.0
prompt-toolkit==2.0.8
protobuf==3.6.1
ptyprocess==0.6.0
py==1.7.0
Pygments==2.3.1
pyparsing==2.3.0
pyrsistent==0.14.11
pytest==4.0.1
python-dateutil==2.7.5
pytz==2018.7
PyWavelets==1.0.1
PyYAML==3.13
pyzmq==18.0.0
qtconsole==4.4.3
scikit-image==0.14.0
scipy==1.1.0
Send2Trash==1.5.0
six==1.11.0
tensorboard==1.12.0
tensorflow==1.12.0
termcolor==1.1.0
terminado==0.8.2
testpath==0.4.2
toolz==0.9.0
tornado==5.1.1
tqdm==4.31.1
traitlets==4.3.2
wcwidth==0.1.7
webencodings==0.5.1
Werkzeug==0.14.1
widgetsnbextension==3.4.2
